+++ app/Console/Kernel.php
@@ -2,9 +2,10 @@
 
 namespace App\Console;
 
-use App\Http\Controllers\LessonOrderController;
+use App\Http\Services\LessonOrderService;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
+use Illuminate\Support\Facades\Log;
 
 class Kernel extends ConsoleKernel
 {
@@ -17,8 +18,10 @@ class Kernel extends ConsoleKernel
     protected function schedule(Schedule $schedule)
     {
         $schedule->call(function () {
-            (new LessonOrderController)->populateOrdersFromFilemaker();
-        })->daily();
+            (new LessonOrderService)->populateOrdersFromFilemaker();
+        })->daily()->onFailure(function () {
+            Log::alert("Populate Orders from Filemaker failed");
+        });
     }
 
     /**