+++ database/migrations/2022_12_14_083707_create_settings_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up()
     {
         Schema::create('settings', function (Blueprint $table): void {
@@ -20,5 +19,9 @@ return new class extends Migration
 
             $table->unique(['group', 'name']);
         });
+    }
+    public function down()
+    {
+        Schema::dropIfExists('settings');
     }
 };