+++ database/migrations/2024_03_11_120447_create_curricula_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->id();
             $table->timestamps();
             $table->string('name');
-            $table->string('email');
+            $table->string('email')->nullable();
             $table->string('jan_lesson')->nullable();
             $table->string('feb_lesson')->nullable();
             $table->string('mar_lesson')->nullable();
@@ -32,6 +32,25 @@ return new class extends Migration {
             $table->string('curriculum_type')->nullable();
             $table->string('grade')->nullable();
         });
+
+        DB::table('curricula')->insert([
+            array(
+                'name' => 'Paper Only',
+                'curriculum_type' => 'paper',
+                'jan_lesson' => 'paper',
+                'feb_lesson' => 'paper',
+                'mar_lesson' => 'paper',
+                'apr_lesson' => 'paper',
+                'may_lesson' => 'paper',
+                'jun_lesson' => 'paper',
+                'jul_lesson' => 'paper',
+                'aug_lesson' => 'paper',
+                'sep_lesson' => 'paper',
+                'oct_lesson' => 'paper',
+                'nov_lesson' => 'paper',
+                'dec_lesson' => 'paper'
+            )
+        ]);
     }
 
     /**